+++ shared/src/commonMain/kotlin/link/limecode/reddit/lite/data/model/ApiPostAttachment.kt
@@ -5,9 +5,9 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ApiPostAttachment(
-    val id: Int,
+    val id: Long,
     @SerialName("post_id")
-    val postId: Int,
+    val postId: Long,
     val name: String,
     @SerialName("attachment_type")
     val attachmentType: ApiAttachmentType
