+++ shared/src/commonMain/kotlin/link/limecode/reddit/lite/data/model/request/subreddit/ApiReqJoinSubReddit.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ApiReqJoinSubReddit(
     @SerialName("subreddit_id")
-    val subRedditId: Int
+    val subRedditId: Long
 )
