+++ shared/src/commonMain/kotlin/link/limecode/reddit/lite/data/model/response/post/ApiResPostList.kt
@@ -6,6 +6,6 @@ import kotlinx.serialization.Serializable
 data class ApiResPostList(
     val items: List<ApiResPostListItem>,
     // TODO: cursor should be long
-    val cursor: Int?,
+    val cursor: Long?,
     val limit: Long
 )
